DB Filesystem: Move all data column related code to `data_columns.go`
Only code move.
